#15: abstract the Project list on the homepage into it's own component and add Project page
